@@ -1,7 +1,8 @@
 package com.pddstudio.highlightjs.demo;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.pddstudio.highlightjs.demo.fragments.FilesListFragment;
 
