@@ -1,10 +1,12 @@
 package com.pddstudio.highlightjs.demo.adapters;
 
-import android.support.v7.widget.RecyclerView;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.pddstudio.highlightjs.demo.R;
 import com.pddstudio.highlightjs.demo.utils.FileObject;
